# [usm] Refactor Kafka encoder