--- a/pkg/network/encoding/kafka.go
+++ b/pkg/network/encoding/kafka.go
@@ -6,64 +6,29 @@
 package encoding
 
 import (
+	"sync"
+
 	model "github.com/DataDog/agent-payload/v5/process"
-	"github.com/cihub/seelog"
+	"github.com/gogo/protobuf/proto"
 
 	"github.com/DataDog/datadog-agent/pkg/network"
+	"github.com/DataDog/datadog-agent/pkg/network/protocols/kafka"
 	"github.com/DataDog/datadog-agent/pkg/network/types"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
+var kafkaAggregationPool = sync.Pool{
+	New: func() any {
+		return &model.KafkaAggregation{
+			Header: new(model.KafkaRequestHeader),
+		}
+	},
+}
+
 type kafkaEncoder struct {
-	aggregations  map[types.ConnectionKey]*kafkaAggregationWrapper
-	orphanEntries int
-}
+	byConnection *USMConnectionIndex[kafka.Key, *kafka.RequestStat]
 
-// aggregationWrapper is meant to handle collision scenarios where multiple
-// `ConnectionStats` objects may claim the same `DataStreamsAggregations` object because
-// they generate the same connection key
-// TODO: we should probably revist/get rid of this if we ever replace socket
-// filters by kprobes, since in that case we would have access to PIDs, and
-// could incorporate that information in the `types.ConnectionKey` struct.
-type kafkaAggregationWrapper struct {
-	*model.DataStreamsAggregations
-
-	// we keep track of the source and destination ports of the first
-	// `ConnectionStats` to claim this `DataStreamsAggregations` object
-	sport, dport uint16
-}
-
-func (a *kafkaAggregationWrapper) ValueFor(c network.ConnectionStats) *model.DataStreamsAggregations {
-	if a == nil {
-		return nil
-	}
-
-	if a.sport == 0 && a.dport == 0 {
-		// This is the first time a ConnectionStats claim this aggregation. In
-		// this case we return the value and save the source and destination
-		// ports
-		a.sport = c.SPort
-		a.dport = c.DPort
-		return a.DataStreamsAggregations
-	}
-
-	if c.SPort == a.dport && c.DPort == a.sport {
-		// We have a collision with another `ConnectionStats`, but this is a
-		// legit scenario where we're dealing with the opposite ends of the
-		// same connection, which means both server and client are in the same host.
-		// In this particular case it is correct to have both connections
-		// (client:server and server:client) referencing the same Kafka data.
-		return a.DataStreamsAggregations
-	}
-
-	// Return nil otherwise. This is to prevent multiple `ConnectionStats` with
-	// exactly the same source and destination addresses but different PIDs to
-	// "bind" to the same DataStreamsAggregations object, which would result in a
-	// overcount problem. (Note that this is due to the fact that
-	// `types.ConnectionKey` doesn't have a PID field.) This happens mostly in the
-	// context of pre-fork web servers, where multiple worker proceses share the
-	// same socket
-	return nil
+	// cached object
+	aggregations *model.DataStreamsAggregations
 }
 
 func newKafkaEncoder(payload *network.Connections) *kafkaEncoder {
@@ -71,75 +36,83 @@ func newKafkaEncoder(payload *network.Connections) *kafkaEncoder {
 		return nil
 	}
 
-	encoder := &kafkaEncoder{
-		aggregations: make(map[types.ConnectionKey]*kafkaAggregationWrapper, len(payload.Conns)),
+	return &kafkaEncoder{
+		aggregations: &model.DataStreamsAggregations{
+			// It's not that important to get the initial size of this slice
+			// right because we're re-using it multiple times and should quickly
+			// converge to a final size after a few calls to
+			// `GetKafkaAggregations`
+			KafkaAggregations: make([]*model.KafkaAggregation, 0, 10),
+		},
+		byConnection: GroupByConnection("kafka", payload.Kafka, func(key kafka.Key) types.ConnectionKey {
+			return key.ConnectionKey
+		}),
 	}
-
-	// pre-populate aggregation map with keys for all existent connections
-	// this allows us to skip encoding orphan Kafka objects that can't be matched to a connection
-	for _, conn := range payload.Conns {
-		for _, key := range network.ConnectionKeysFromConnectionStats(conn) {
-			if log.ShouldLog(seelog.TraceLvl) {
-				log.Tracef("Payload has a connection %v and was converted to kafka key %v", conn, key)
-			}
-			encoder.aggregations[key] = nil
-		}
-	}
-	encoder.buildAggregations(payload)
-	return encoder
 }
 
-func (e *kafkaEncoder) GetKafkaAggregations(c network.ConnectionStats) *model.DataStreamsAggregations {
+func (e *kafkaEncoder) GetKafkaAggregations(c network.ConnectionStats) []byte {
 	if e == nil {
 		return nil
 	}
 
-	connectionKeys := network.ConnectionKeysFromConnectionStats(c)
-	for _, key := range connectionKeys {
-		if aggregation := e.aggregations[key]; aggregation != nil {
-			return e.aggregations[key].ValueFor(c)
-		}
+	connectionData := e.byConnection.Find(c)
+	if connectionData == nil || len(connectionData.Data) == 0 || connectionData.IsPIDCollision(c) {
+		return nil
 	}
-	return nil
+
+	return e.encodeData(connectionData)
 }
 
-func (e *kafkaEncoder) buildAggregations(payload *network.Connections) {
-	// Count the number of aggregations per connections, so we can set the capacity of the aggregation slice accordingly
-	aggregationsCountPerConnection := make(map[types.ConnectionKey]int)
-	for key := range payload.Kafka {
-		aggregationsCountPerConnection[key.ConnectionKey]++
+func (e *kafkaEncoder) Close() {
+	if e == nil {
+		return
 	}
 
-	for key, stats := range payload.Kafka {
-		aggregation, ok := e.aggregations[key.ConnectionKey]
-		if !ok {
-			// if there is no matching connection don't even bother to serialize Kafka data
-			if log.ShouldLog(seelog.TraceLvl) {
-				log.Tracef("Found kafka orphan connection %v", key.ConnectionKey)
-			}
-			e.orphanEntries++
-			continue
-		}
+	e.reset()
+	e.byConnection.Close()
+}
 
-		if aggregation == nil {
-			// No aggregation created for this connection yet, creating the slice and set the capacity accordingly
-			aggregation = &kafkaAggregationWrapper{
-				DataStreamsAggregations: &model.DataStreamsAggregations{
-					KafkaAggregations: make([]*model.KafkaAggregation, 0, aggregationsCountPerConnection[key.ConnectionKey]),
-				},
-			}
-			e.aggregations[key.ConnectionKey] = aggregation
-		}
+func (e *kafkaEncoder) encodeData(connectionData *USMConnectionData[kafka.Key, *kafka.RequestStat]) []byte {
+	e.reset()
 
-		kafkaAggregation := model.KafkaAggregation{
-			Header: &model.KafkaRequestHeader{
-				RequestType:    uint32(key.RequestAPIKey),
-				RequestVersion: uint32(key.RequestVersion),
-			},
-			Topic: key.TopicName,
-			Count: uint32(stats.Count),
-		}
+	for _, kv := range connectionData.Data {
+		key := kv.Key
+		stats := kv.Value
 
-		aggregation.KafkaAggregations = append(aggregation.KafkaAggregations, &kafkaAggregation)
+		kafkaAggregation := kafkaAggregationPool.Get().(*model.KafkaAggregation)
+
+		kafkaAggregation.Header.RequestType = uint32(key.RequestAPIKey)
+		kafkaAggregation.Header.RequestVersion = uint32(key.RequestVersion)
+		kafkaAggregation.Topic = key.TopicName
+		kafkaAggregation.Count = uint32(stats.Count)
+
+		e.aggregations.KafkaAggregations = append(e.aggregations.KafkaAggregations, kafkaAggregation)
 	}
+
+	serializedData, _ := proto.Marshal(e.aggregations)
+	return serializedData
+}
+
+func (e *kafkaEncoder) reset() {
+	if e == nil {
+		return
+	}
+
+	for i, kafkaAggregation := range e.aggregations.KafkaAggregations {
+		// The pooled *model.KafkaAggregation object comes along with a
+		// pre-allocated *model.KafkaHeader object as well, so we ensure that we
+		// clean both objects but keep them linked together before returning it
+		// to the pool.
+
+		header := kafkaAggregation.Header
+		header.Reset()
+
+		kafkaAggregation.Reset()
+		kafkaAggregation.Header = header
+
+		kafkaAggregationPool.Put(kafkaAggregation)
+		e.aggregations.KafkaAggregations[i] = nil
+	}
+
+	e.aggregations.KafkaAggregations = e.aggregations.KafkaAggregations[:0]
 }