--- a/pkg/network/encoding/kafka.go
+++ b/pkg/network/encoding/kafka.go
@@ -73,8 +73,6 @@ func (e *kafkaEncoder) Close() {
 }
 
 func (e *kafkaEncoder) encodeData(connectionData *USMConnectionData[kafka.Key, *kafka.RequestStat]) []byte {
-	e.reset()
-
 	for _, kv := range connectionData.Data {
 		key := kv.Key
 		stats := kv.Value