--- a/pkg/network/protocols/kafka/kafka_stats.go
+++ b/pkg/network/protocols/kafka/kafka_stats.go
@@ -24,10 +24,12 @@ type Key struct {
 }
 
 // NewKey generates a new Key
-func NewKey(saddr, daddr util.Address, sport, dport uint16, topicName string) Key {
+func NewKey(saddr, daddr util.Address, sport, dport uint16, topicName string, requestAPIKey, requestAPIVersion uint16) Key {
 	return Key{
-		ConnectionKey: types.NewConnectionKey(saddr, daddr, sport, dport),
-		TopicName:     topicName,
+		ConnectionKey:  types.NewConnectionKey(saddr, daddr, sport, dport),
+		TopicName:      topicName,
+		RequestAPIKey:  requestAPIKey,
+		RequestVersion: requestAPIVersion,
 	}
 }