<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Draft PRs should be prefixed with `[WIP]` in their title.

-->
### What does this PR do?
Refactor the Kafka encoder by leveraging changes introduced in https://github.com/DataDog/datadog-agent/pull/16844
Optimize USM Kafka encoding codepath.
The main idea of this PR is reducing the number of live protobuf `model` objects in the heap at given time. The approach can be summarized as follows:

1. First we group USM data by connection calling `GroupByConnection`;
2. Then, during the encoding phase of each `network.ConnectionStats` we do the following:
  a. First we fetch all USM data for a given connection (which is done by calling `USMConnectionIndex.Find()`)
  b. Then, we generate all Kafka `model` objects for this specific connection
  c. After that, we serialize all objects into a blob of bytes
  d. The associated `model` objects are then moved to an object pool for later reuse.
 
<!-- In aggregate the changes in this PR have yielded a ~20% reduction in RSS and a ~10% reduction in CPU in our load test environment running an Kafka-heavy workload.
 
![rss](https://user-images.githubusercontent.com/692520/235477409-49168161-67be-4353-9217-8b341984ef39.png)
![cpu](https://user-images.githubusercontent.com/692520/235477450-44f34e32-32a8-4212-9990-7f951137b30f.png)
-->

### Motivation

<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

Benchmark on the last commit:

```
Before
BenchmarkKafkaEncoder100Requests
BenchmarkKafkaEncoder100Requests-16      	  113868	     10180 ns/op	   12140 B/op	      22 allocs/op
BenchmarkKafkaEncoder10000Requests
BenchmarkKafkaEncoder10000Requests-16    	    1287	    926992 ns/op	 1164124 B/op	      24 allocs/op

After
BenchmarkKafkaEncoder100Requests
BenchmarkKafkaEncoder100Requests-16      	  133870	      8973 ns/op	    9344 B/op	      22 allocs/op
BenchmarkKafkaEncoder10000Requests
BenchmarkKafkaEncoder10000Requests-16    	    1639	    688846 ns/op	  722081 B/op	      23 allocs/op
```

Load test:
A moderate change in avg CPU, up to 10%
![image](https://github.com/DataDog/datadog-agent/assets/17148247/901556df-8dd7-45a6-9344-d6dd34305ad3)

Max CPU seems stable with main
![image](https://github.com/DataDog/datadog-agent/assets/17148247/0916e21f-7db6-4120-90fd-9d02fdce6eaf)

A slight improvement in AVG RSS
![image](https://github.com/DataDog/datadog-agent/assets/17148247/a63c45f3-5dd1-4449-9b71-525c2898532b)


Max RSS looks stable
![image](https://github.com/DataDog/datadog-agent/assets/17148247/315ac638-df58-4b0a-a58b-1c3fc8c55097)

Reduction in working set
![image](https://github.com/DataDog/datadog-agent/assets/17148247/b430ce9c-98c4-473d-9484-3f930f959c09)

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs

<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes

<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature. 
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.