@guyarb 2023-06-17
> Before we merge this, is there a load testing cluster we can trial this on ?

I tested in a load test, the results appear in the PR description